requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
numpy==2.3.4
pillow==12.0.0
qrcode==8.2
scipy==1.16.3
setuptools==80.9.0
Werkzeug==3.1.3
wheel==0.45.1
